- Cleaned up unneeded zcap code (RootCapability type)
